+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Airbrake Ruby Changelog
 
 * Added support for custom exception attributes
   ([#113](https://github.com/airbrake/airbrake-ruby/pull/113))
+* Started validating the 'environment' config option (a warning will be printed,
+  if it is misconfigured)
+  ([#115](https://github.com/airbrake/airbrake-ruby/pull/115))
 
 ### [v1.4.6][v1.4.6] (August 18, 2016)